Move GenericCommandState path substitution in substitution.rs
